flask_app/requirements.txt
flask
flask-cors
requests
python-dotenv
pyjwt
passlib[bcrypt]
werkzeug
